+++ api/questions.js
@@ -7,8 +7,14 @@ const { matchedData } = require('express-validator/filter')
 
 const constants = require('../constants')
 const { Course, Queue, Question } = require('../models/')
-const { requireQueue, requireQuestion, failIfErrors } = require('./util')
+const {
+  requireQueue,
+  requireQueueForQuestion,
+  requireQuestion,
+  failIfErrors,
+} = require('./util')
 const requireCourseStaffForQueueForQuestion = require('../middleware/requireCourseStaffForQueueForQuestion')
+const safeAsync = require('../middleware/safeAsync')
 
 /* eslint-disable no-param-reassign */
 function modifyBeingAnswered(question, answering) {
@@ -21,6 +27,19 @@ function modifyBeingAnswered(question, answering) {
 }
 /* eslint-enable no-param-reassign */
 
+function checkLocation(req, res, next) {
+  check('location')
+    .custom(value => {
+      if (res.locals.queue.fixedLocation) return true
+      return (
+        !!value &&
+        value.length > 0 &&
+        value.length <= constants.QUESTION_LOCATION_MAX_LENGTH
+      )
+    })
+    .trim()(req, res, next)
+}
+
 // Adds a question to a queue
 router.post(
   '/',
@@ -32,12 +51,10 @@ router.post(
     check('topic')
       .isLength({ min: 1, max: constants.QUESTION_TOPIC_MAX_LENGTH })
       .trim(),
-    check('location')
-      .isLength({ min: 1, max: constants.QUESTION_LOCATION_MAX_LENGTH })
-      .trim(),
+    checkLocation,
     failIfErrors,
   ],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const data = matchedData(req)
     const { id: queueId } = res.locals.queue
 
@@ -56,7 +73,8 @@ router.post(
 
     const question = Question.build({
       name: data.name,
-      location: data.location,
+      // Questions in fixed-location queues should never have a location
+      location: res.locals.queue.fixedLocation ? '' : data.location,
       topic: data.topic,
       enqueueTime: new Date(),
       queueId,
@@ -66,21 +84,25 @@ router.post(
     await question.save()
     await question.reload()
     res.status(201).send(question)
-  }
+  })
 )
 
 // Get all questions for a particular queue
-router.get('/', [requireQueue, failIfErrors], async (req, res, _next) => {
-  const { id: queueId } = res.locals.queue
-  const questions = await Question.findAll({
-    where: {
-      queueId,
-      dequeueTime: null,
-    },
-    order: [['id', 'ASC']],
+router.get(
+  '/',
+  [requireQueue, failIfErrors],
+  safeAsync(async (req, res, _next) => {
+    const { id: queueId } = res.locals.queue
+    const questions = await Question.findAll({
+      where: {
+        queueId,
+        dequeueTime: null,
+      },
+      order: [['id', 'ASC']],
+    })
+    res.send(questions)
   })
-  res.send(questions)
-})
+)
 
 router.get(
   '/:questionId',
@@ -93,27 +115,55 @@ router.get(
 // Mark a question as being answered
 router.post(
   '/:questionId/answering',
-  [requireCourseStaffForQueueForQuestion, requireQuestion, failIfErrors],
-  async (req, res, _next) => {
-    const { question } = res.locals
+  [
+    requireCourseStaffForQueueForQuestion,
+    requireQuestion,
+    requireQueueForQuestion,
+    failIfErrors,
+  ],
+  safeAsync(async (req, res, _next) => {
+    const { queue, question } = res.locals
+
+    if (question.beingAnswered) {
+      // Forbid someone else from taking over this question
+      res.status(403).send('Another user is already answering this question')
+      return
+    }
+
+    // Verify that this user isn't currently answering another question
+    const otherQuestions = await Question.find({
+      where: {
+        answeredById: res.locals.userAuthn.id,
+        dequeueTime: null,
+        queueId: queue.id,
+      },
+    })
+
+    if (otherQuestions !== null) {
+      res
+        .status(403)
+        .send('You are already answering another question on this queue')
+      return
+    }
+
     modifyBeingAnswered(question, true)
     question.answeredById = res.locals.userAuthn.id
     await question.save()
     res.send(question)
-  }
+  })
 )
 
 // Mark a question as no longer being answered
 router.delete(
   '/:questionId/answering',
   [requireCourseStaffForQueueForQuestion, requireQuestion, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const { question } = res.locals
     modifyBeingAnswered(question, false)
     question.answeredById = null
     await question.save()
     res.send(question)
-  }
+  })
 )
 
 // Mark the question as answered
@@ -128,48 +178,61 @@ router.post(
       .trim(),
     failIfErrors,
   ],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const data = matchedData(req)
+
+    // Temporary, easy fix to avoid having to rename enums
+    // TODO Fix this garbage
+    let mappedPreparedness = data.preparedness
+    switch (data.preparedness) {
+      case 'bad':
+        mappedPreparedness = 'not'
+        break
+      case 'good':
+        mappedPreparedness = 'well'
+        break
+      default:
+        break
+    }
 
     const { question } = res.locals
     question.answerFinishTime = new Date()
     question.dequeueTime = new Date()
-    question.preparedness = data.preparedness
+    question.preparedness = mappedPreparedness
     question.comments = data.comments
     question.answeredById = res.locals.userAuthn.id
 
     const updatedQuestion = await question.save()
     res.send(updatedQuestion)
-  }
+  })
 )
 
-// updates a question's location
+// Updates a question's information
 router.patch(
   '/:questionId',
   [
     requireQuestion,
-    check('location')
-      .isLength({ min: 1, max: constants.QUESTION_LOCATION_MAX_LENGTH })
-      .trim(),
+    requireQueueForQuestion,
     check('topic')
       .isLength({ min: 1, max: constants.QUESTION_TOPIC_MAX_LENGTH })
       .trim(),
+    checkLocation,
     failIfErrors,
   ],
-  async (req, res, _next) => {
-    const { userAuthn, question } = res.locals
+  safeAsync(async (req, res, _next) => {
+    const { userAuthn, question, queue } = res.locals
     const data = matchedData(req)
 
     if (question.askedById === userAuthn.id) {
       await question.update({
-        location: data.location,
+        location: queue.fixedLocation ? '' : data.location,
         topic: data.topic,
       })
       res.status(201).send(question)
     } else {
       res.status(403).send()
     }
-  }
+  })
 )
 
 // Deletes a question from a queue, without marking
@@ -178,7 +241,7 @@ router.patch(
 router.delete(
   '/:questionId',
   [requireQuestion, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const { userAuthn, userAuthz, question } = res.locals
     const { queueId } = question
 
@@ -205,7 +268,7 @@ router.delete(
     } else {
       res.status(403).send()
     }
-  }
+  })
 )
 
 module.exports = router