+++ api/queues.js
@@ -2,10 +2,11 @@ const router = require('express').Router({
   mergeParams: true,
 })
 
-const { check } = require('express-validator/check')
+const { check, oneOf } = require('express-validator/check')
 const { matchedData } = require('express-validator/filter')
 
 const { Queue, ActiveStaff, Question, User } = require('../models')
+const safeAsync = require('../middleware/safeAsync')
 
 const {
   requireCourse,
@@ -17,11 +18,21 @@ const {
 const requireCourseStaffForQueue = require('../middleware/requireCourseStaffForQueue')
 const requireCourseStaff = require('../middleware/requireCourseStaff')
 
+function validateLocation(req, res, next) {
+  check('location').custom(value => {
+    if (!res.locals.queue.fixedLocation) return true
+    return !!value && value.length > 0
+  })(req, res, next)
+}
+
 // Get all open queues
-router.get('/', async (req, res, _next) => {
-  const queues = await Queue.scope('questionCount').findAll()
-  res.json(queues)
-})
+router.get(
+  '/',
+  safeAsync(async (req, res, _next) => {
+    const queues = await Queue.scope('questionCount').findAll()
+    res.json(queues)
+  })
+)
 
 // Create a queue for a course
 router.post(
@@ -30,31 +41,43 @@ router.post(
     requireCourseStaff,
     requireCourse,
     check('name').isLength({ min: 1 }),
-    check('location').optional({ nullable: true }),
+    oneOf([
+      [
+        check('fixedLocation').custom(value => value !== true),
+        check('location').optional({ nullable: true }),
+      ],
+      [
+        check('fixedLocation').custom(value => value === true),
+        check('location').isLength({ min: 1 }),
+      ],
+    ]),
     failIfErrors,
   ],
-  (req, res, _next) => {
+  safeAsync(async (req, res, next) => {
     const { id: courseId } = res.locals.course
     const data = matchedData(req)
 
-    const queue = Queue.build({
+    const queue = await Queue.create({
       name: data.name,
       location: data.location,
+      fixedLocation: data.fixedLocation === true,
       courseId,
       createdByUserId: res.locals.userAuthn.id,
     })
 
-    queue.save().then(newQueue => res.status(201).json(newQueue))
-  }
+    Queue.scope('questionCount')
+      .findById(queue.id)
+      .then(newQueue => res.status(201).json(newQueue))
+      .catch(next)
+  })
 )
 
 // Gets a specific queue
 router.get(
   '/:queueId',
   [requireQueue, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const { id: queueId } = res.locals.queue
-
     const queue = await Queue.findOne({
       where: {
         id: queueId,
@@ -79,15 +102,41 @@ router.get(
       ],
       order: [[Question, 'id', 'ASC']],
     })
+
     res.json(queue)
-  }
+  })
+)
+
+// Modify a queue for a course
+router.patch(
+  '/:queueId',
+  [
+    requireCourseStaffForQueue,
+    requireQueue,
+    check('name').isLength({ min: 1 }),
+    validateLocation,
+    failIfErrors,
+  ],
+  safeAsync(async (req, res, _next) => {
+    const { queue } = res.locals
+    const data = matchedData(req)
+
+    await queue.update({
+      name: data.name,
+      location: data.location,
+    })
+    const updatedQueue = await Queue.scope('questionCount').findOne({
+      where: { id: queue.id },
+    })
+    res.status(201).send(updatedQueue)
+  })
 )
 
 // Gets the on-duty staff list for a specific queue
 router.get(
   '/:queueId/staff',
   [requireQueue, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const { queue } = res.locals
     const staff = await ActiveStaff.findAll({
       where: {
@@ -97,14 +146,14 @@ router.get(
       include: [User],
     })
     res.json(staff)
-  }
+  })
 )
 
 // Joins the specified user to the specified queue
 router.post(
   '/:queueId/staff/:userId',
   [requireCourseStaffForQueue, requireQueue, requireUser, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const { id: userId } = res.locals.userAuthn
     const { id: queueId } = res.locals.queue
     const [staff, created] = await ActiveStaff.findOrCreate({
@@ -132,14 +181,14 @@ router.post(
     } else {
       res.status(202).json(staff)
     }
-  }
+  })
 )
 
 // Removes the specified user from the specified queue
 router.delete(
   '/:queueId/staff/:userId',
   [requireCourseStaffForQueue, requireQueue, requireUser, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     const { id: userId } = res.locals.user
     const { id: queueId } = res.locals.queue
     const staff = await ActiveStaff.find({
@@ -156,17 +205,17 @@ router.delete(
     }
 
     res.status(202).send()
-  }
+  })
 )
 
 // Deletes the queue
 router.delete(
   '/:queueId',
   [requireCourseStaffForQueue, requireQueue, failIfErrors],
-  async (req, res, _next) => {
+  safeAsync(async (req, res, _next) => {
     await res.locals.queue.destroy()
     res.status(202).send()
-  }
+  })
 )
 
 module.exports = router