+++ components/QuestionList.js
@@ -150,6 +150,9 @@ class QuestionList extends React.Component {
               isUserAnsweringQuestion={
                 this.props.userId === question.answeredById
               }
+              isUserAnsweringOtherQuestion={
+                this.props.isUserAnsweringQuestionForQueue
+              }
               didUserAskQuestion={this.props.userId === question.askedById}
               deleteQuestion={() => this.deleteQuestion(questionId)}
               cancelQuestion={() => this.cancelQuestion(questionId)}
@@ -213,6 +216,7 @@ class QuestionList extends React.Component {
         />
         <QuestionEdit
           question={this.props.questions[this.state.attributeId]}
+          queue={this.props.queue}
           isOpen={this.state.showQuestionEditModal}
           onSubmitQuestionEdit={attributes =>
             this.handleSubmitQuestionEdit(attributes)
@@ -227,6 +231,7 @@ class QuestionList extends React.Component {
 QuestionList.propTypes = {
   queue: PropTypes.shape({
     questions: PropTypes.arrayOf(PropTypes.number),
+    fixedLocation: PropTypes.bool,
   }),
   questions: PropTypes.objectOf(
     PropTypes.shape({
@@ -237,6 +242,7 @@ QuestionList.propTypes = {
   ),
   userId: PropTypes.number.isRequired,
   isUserCourseStaff: PropTypes.bool.isRequired,
+  isUserAnsweringQuestionForQueue: PropTypes.bool.isRequired,
   deleteQuestion: PropTypes.func.isRequired,
   updateQuestionBeingAnswered: PropTypes.func.isRequired,
   finishAnsweringQuestion: PropTypes.func.isRequired,